Move out from tests to main
